fixes in refactoring
