+++ iidm/iidm-xml-converter/src/main/java/com/powsybl/iidm/xml/NetworkXml.java
@@ -18,6 +18,7 @@ import com.powsybl.commons.xml.XmlUtil;
 import com.powsybl.iidm.anonymizer.Anonymizer;
 import com.powsybl.iidm.anonymizer.SimpleAnonymizer;
 import com.powsybl.iidm.export.BusFilter;
+import com.powsybl.iidm.export.ExportOptions;
 import com.powsybl.iidm.import_.ImportOptions;
 import com.powsybl.iidm.network.*;
 import javanet.staxutils.IndentingXMLStreamWriter;
@@ -71,7 +72,7 @@ public final class NetworkXml {
     private NetworkXml() {
     }
 
-    private static XMLStreamWriter createXmlStreamWriter(XMLExportOptions options, OutputStream os) throws XMLStreamException {
+    private static XMLStreamWriter createXmlStreamWriter(ExportOptions options, OutputStream os) throws XMLStreamException {
         XMLStreamWriter writer = XML_OUTPUT_FACTORY_SUPPLIER.get().createXMLStreamWriter(os, StandardCharsets.UTF_8.toString());
         if (options.isIndent()) {
             IndentingXMLStreamWriter indentingWriter = new IndentingXMLStreamWriter(writer);
@@ -190,7 +191,7 @@ public final class NetworkXml {
         }
     }
 
-    public static Anonymizer write(Network n, XMLExportOptions options, OutputStream os) {
+    public static Anonymizer write(Network n, ExportOptions options, OutputStream os) {
         try {
             final XMLStreamWriter writer = createXmlStreamWriter(options, os);
             writer.writeStartDocument(StandardCharsets.UTF_8.toString(), "1.0");
@@ -246,10 +247,10 @@ public final class NetworkXml {
     }
 
     public static Anonymizer write(Network n, OutputStream os) {
-        return write(n, new XMLExportOptions(), os);
+        return write(n, new ExportOptions(), os);
     }
 
-    public static Anonymizer write(Network n, XMLExportOptions options, Path xmlFile) {
+    public static Anonymizer write(Network n, ExportOptions options, Path xmlFile) {
         try (OutputStream os = new BufferedOutputStream(Files.newOutputStream(xmlFile))) {
             return write(n, options, os);
         } catch (IOException e) {
@@ -258,7 +259,7 @@ public final class NetworkXml {
     }
 
     public static Anonymizer write(Network n, Path xmlFile) {
-        return write(n, new XMLExportOptions(), xmlFile);
+        return write(n, new ExportOptions(), xmlFile);
     }
 
     public static Anonymizer writeAndValidate(Network n, Path xmlFile) {
