+++ iidm/iidm-xml-converter/src/test/java/com/powsybl/iidm/xml/TopologyLevelTest.java
@@ -9,6 +9,7 @@ package com.powsybl.iidm.xml;
 import com.powsybl.commons.AbstractConverterTest;
 import com.powsybl.commons.config.InMemoryPlatformConfig;
 import com.powsybl.commons.config.PlatformConfig;
+import com.powsybl.iidm.export.ExportOptions;
 import com.powsybl.iidm.network.*;
 import com.powsybl.iidm.network.test.FictitiousSwitchFactory;
 import org.junit.Before;
@@ -57,21 +58,21 @@ public class TopologyLevelTest extends AbstractConverterTest {
     }
 
     private static void writeNodeBreaker(Network network, Path path) {
-        XMLExportOptions options = (XMLExportOptions) new XMLExportOptions()
+        ExportOptions options = new ExportOptions()
             .setTopologyLevel(TopologyLevel.NODE_BREAKER);
 
         NetworkXml.write(network, options, path);
     }
 
     private static void writeBusBreaker(Network network, Path path) {
-        XMLExportOptions options = (XMLExportOptions) new XMLExportOptions()
+        ExportOptions options = new ExportOptions()
             .setTopologyLevel(TopologyLevel.BUS_BREAKER);
 
         NetworkXml.write(network, options, path);
     }
 
     private static void writeBusBranch(Network network, Path path) {
-        XMLExportOptions options = (XMLExportOptions) new XMLExportOptions()
+        ExportOptions options = new ExportOptions()
             .setTopologyLevel(TopologyLevel.BUS_BRANCH);
 
         NetworkXml.write(network, options, path);
